Downgrade Next.js version to avoid broken build
We have to figure out a way to stop using jsonwebtoken inside _middleware. Functions defined in there are to be run on the Edge, which adds extra limitations in what's available or not. The alternative is to use @tsndr/cloudflare-worker-jwt, however, the change requires more work than simply switching imports.