Wrap button within widget component
We might always use this component in the same way and the button needs to be wrapped in a NextLink.

Therefore, to avoid importing all those pieces inside the caller, I'm moving the button to PermissionDeniedWidget and turning the children prop into a description text added to the message at will.